Add govuk image assets into CopyWebpack pattern
Govuk browser tab icons are being used for apple in DfE design manual so we will use the same pattern. These need to be manually copied into dist so that they can access in html layout template.

We may want to review the assets path in the application stylesheet to prevent duplication.