Change trust search to use Academies API
Currently the Academies API `/trusts` endpoint is only returning 50 results, which do not have much data. This is holding up development of the search results page. We will temporarily use the `/trusts` endpoint directly to search, by passing the `groupName` query term.
For this, we have added an additional method to the `TrustProvider` class and will not use the `TrustSearch` as we are not handling the search in the application.

We are intending to revert to search in the app, to enable us to cache trusts on which to search-in which case we will reinstate the `TrustSearch` class to handle this.

Replace Search class with provider class