Move mock setup to test dependency
As we only want to register mock dependencies, and only for tests run in the isolated CI environment, we have moved the mock configuration into a setup task.

We have also created an isolated config for tests run in the CI environment, as this way we can add the dependency to all browser projects in this configuration.