Remove GOV.UK heading style
The use of a GOV.UK heading style class on a link element makes the focus area of the link spread across the width of the page. This causes difficulties for a user trying to highlight and copy the trust name. As this element is not actually a heading, have removed this class and replicated with styles to override the font size, weight and spacing.