Extract search box input to partial
This in preparation for adding the search box to the Search page. The Search box will have the same appearance and behaviour in both places, so can be reused.

Added a new search model to pass to the partial to manage the data binding for the form, and also passing the input id as a ViewData variable, in order to allow different labels for the input to be added at the parent page level.