Update mock trust provider with query matching
Our mock trust provider was previously returning one canned response, no matter what query parameter was provided through search. This was limiting our UI tests:
- We were unable to check the right results were displayed if a user searched more than once
- We would not be able to test the UI for if no results were returned

This change adds a fixed number of query matched responses, with different lengths. The page object model has also been updated to store the current search term - so that we do not need to let it know which term we are checking for in each of our assertions.