Replace security headers in middleware with NetEscapades
The application requires the use of inline script tags, which we need to secure with nonces. We have added nonce attributes using the NetEscapade package, as commonly used in RSD products.

In addition we have made further use of NetEscapade security header methods to replace the majority of security headers previously set in our middleware - so that all of the security headers are set in one place. We took this opportunity to review which headers are being set, and made some small amendments.

We still need to manual set the X-Robots-Tag using the custom middleware, to define search engine behaviour.