Add check for matching trust id and keywords
There is a known bug on the autocomplete behaviour - if a user selects and item and changes the text input to trigger a manual search, the selected item will not be deselected.

This change adds a check on the search page, to check the search term matches the name of the trust with the associated selected trust id before navigating to the selected trust page. If the search term does not match a list of results is shown instead.