Add UI tests for autocomplete on search page
Testing that a user can go directly to a trust page when searching, by selecting a suggestion from the dropdown, and that a user can select an item and change their search before clicking search.

Also testing that a user can use the autocomplete from the search page after searching on a different page (e.g. using home page search).