Show no results found message
This adds the autocomplete component property `showNoOptionsFound=true` to display a no results found message when using autocomplete, along with a UI test for the behaviour.

Currently the message is shown briefly before results are returned which could confuse users. This will also mean that the expect assertion for this will be true for each of our tests - but it does test that it is displayed.