Skip accessibility test for no results
Skipping this test as the autocomplete element fails accessibility tests when showing the no results found message

message: 'Element has children which are not allowed'

This is referring to an li element, however the ul and li nesting seems to be correct.

As this is not our application code we will skip this test for now, and see if we face any issues in our audit.