Fix font sizes not changing correctly on autocomplete
Manual testing of this component has found that the font in the input or autcomplete options do not change when the browser font is changed.

Have added the `govuk-font` mixin to set the font-size to 19px using rem.

Have also removed the set widths on the search icon svg, and moved this to scss, so that the icon sizes accordingly.