Fix search button wrapping to new line
We were using the `govuk-input--wrapper` class to style the search input and search button to sit side by side, however this wraps to a new line on smaller screens, and would be hidden by the autocomplete element.
Removed the class, and added the relevant styles to our `app-search--wrapper` class without the media query. We will see how this tests with users using smaller screens and larger font sizes - to check the search input is large enough.