Add trust type to Trust record
TL;DR: For this change, have renamed the `Trust` record in search, and added a new `Trust` record to be used on trust pages.

As trust type is not required for a search results listing, have diverged into two different models or 'record' types for a trust. Search will now use the record type `TrustSearchEntry` to match `AutocompletEntry`, while the `Trust` record will contain more properties as needed for Trust pages.