Install digital scotland component library
Import and initialise component 'Side Navigation' - as this is the only component we plan to use from the library.

Reset style for `ul` - Digital Scotland has set styles for native elements, which has changed the appearance of the search results list on search page. Resetting native list elements using `resets.scss` to fix this.