Add navigation to contacts page
Add contacts page so that we can style navigation between two different nav items on the side navigation.

Contacts page is currently fetching Trust information using the same method as for the details page, but it is expected that we will change this when adding connection to the database, and when completing each of these pages.

Add tests for Contacts page