Move Page title and headings to shared TrustLayout
As each title will include the name of the trust, setting a new ViewData propert 'Heading' in the page, while the title will be a combination of the page heading and the trust name.

Use ViewData to pass the page and section heading to shared `TrustLayout`, so that styles can be consistent.