Override Digital Scotland breakpoints
Digital Scotland breakpoints do not match with the GOV.UK grid system, and so the side navigation is showing as an expandable menu before the grid has collapsed.
We have overridden the breakpoints directly from the $govuk-breakpoints sass map, but provided fallback variables should the mapping change in the future.

Disable stylelint rule for breakpoint override and expand colour, as Digital Scotland naming patterns do not match the guidelines but we cannot change them.