Add visually hidden text on side nav links
A WAVE scan of the trust layout has suggested that there is not enough context to the 'details' link text. Although the context is provided visually by the navigation heading, adding some visually hidden text to make it clear that both 'details' and 'contacts' links are about the trust, rather than the entire site.