Set a max number of retries on CI
As we run a retry with trace on failed tests, we want to ensure that the whole suite is not rerun if all (or a large number of tests) are failing.

We have set this to 20 so that we can see whether the failure is common across tests - as we rerun failed tests with a trace in CI. 20 failures will allow 10 failed tests with retries.

We will set max failures locally too.