Update banner background colour
Matching the background colour of the phase banner to that of the app banner caused our accessibility tests to fail, as there is not a high enough contrast ratio between the link text and the background colour.

Have tweaked the variable for the background colour to match theat used on the DfE Design system design manual (design.education.gov.uk). This is now a high enough contrast ratio, and our accessibility tests are passing.