Reduce Digital Scotland base scss imports
We have installed the Digital Scotland Design system in order to use the Side Navigation component for Trust page layouts.

While we have only imported the component we are using into our scss files, we are also required to import the base styles so that the file has access to Digital Scotland colour variables and media queries.

However, the base styles apply Digital Scotland styles to html tags such as `a` and `ul`, which we do not want in our application (as we are using the GOV.UK design system).

Here, am specifically importing only what is required for the side navigation to work.

We do need the `ds-link` style from the core styles, as this is programmatically applied in the js version of side nav. Unfortunately this is tied to the `a` tag in the stylesheet, so directly applying the style as in the file in our `_digital-scotland.scss` file.