Create new Trust search provider in database project
Create new class in `Dfe.FindInformationAcademiesTrusts.Db` which will handle trust search. This should pass our exisitng tests for `TrustSearch`, so have moved these to `Dfe.FindInformationAcademiesTrusts.Db.UnitTests` with a MockDbContext to mock our connection to the database.

We have changed the tests/implementation for an empty search term to return an empty list, before calling the database.

We will need to add additional tests and implementation for the following:
- Trust search entry should include an address
- Trust search entry should include identifiers