Move left alignment of nav to media query
Digital Scotland's side navigation component has a nesting similar to a list html element. In our design, we are left-aligning the side-navigation, however a side-effect of this is that in the button-first design for small screens (when JavaScript is on), the target area is cropped to the edge of the text on the left. We have agreed with design that we will only left-align the menu on larger screens, and keep the horizontal nesting on the collapsed menu.

To do this, have moved the negative margin to inside the media query for large screens.