Check that GroupID and GroupUID are not null and filter out non trust groups
We only want to return single and multi academy trusts so we need to filter out all other group types including Trust.
GroupUid and GroupId are never null for single and multi academy trusts but we should still check as they are nullable fields and we want to use non nullable C# properties.