Create Faker project
Create new console application `DfE.FindInformationAcademiesTrusts` which we will use to generate seed data for our tests.

We have placed this inside our tests folder as its primary function will be to generate data to use in tests. We may also use this project for seeding data in the local development environment, but we wanted to clearly separate it from production code.