Create GroupFaker
Create `GroupFaker` class based on the Group table model from the `DfE.FindAcademiesTrusts.AcademiesDb` project.

This Faker only generates data for the fields we are currently using in the application.

Emitting the result of the generated data to a json file for now.

This project uses the Fake Data Generator developed for the TRAMS-alpha prototype as a basis, and have taken the relevant example data and helpers from here.

https://github.com/dynamictulip/FakeDataGenerator