county and addressline3 removed from model,dbcontext and address builder
no trusts in the database have a county or a third line address recorded. in the database the missing data is recorded as string of 'not recorded'. this string is being surfaced in our group model which is not needed at this present time. the two properties have been removed from the group model and the dbcontext also.The county field has also been removed from address builder and will no longer surface in the autocompleted searches of trusts.