Replace `GroupUid` with `Uid` outside academies db project
We want to keep our language in the web project consistent with that of our users in order to make the system easier to maintain - we'd started to drift already with `GroupUid`, `Uid` and `TrustId` all meaning the same thing. `GroupUid` is language specific to the academies database so we will keep it there and use `Uid` in all other scenarios.

Also refactored `SearchModelTests`