Add UI tests for trust details page
Add a UI test which checks that the correct trust information is displayed on the trust details page.

As this change is purely UI work, using static data the page object model contains the same static data to use in the tests. Once the data is populated correctly in a later user story this test is expected to fail and the page object model will need to be updated to use the AcademiesDb faker to check the correct values are displaying for different trusts.