Pass database password as argument
Currently the database password is being referred to in several places. We want to declare it only in theDocker Compose steps to the Dockerfile so that we can spot any issues with differences in the database creation and the connection string.

As ENV variables are not available during build time, we will pass it as a build ARG. This can also be passed to the bash `import-data.sh` as an argument to use the password when running each of the sql scripts to create and seed the database.