Simplify spec files for home and search pages
Remove any data such as search term or trust name from spec files, and instead pass a 'current search' object to page object models which shares information on the current search between page object models.

This helps us to remove our reliance on canned data set up in the Mock API, as we have replaced it with a fake database which contains a larger set of data.

We will still need to check specific data in the search reulsts list, and trust information pages, for which we will use a json file generated during the fake data generator step.

Change tests with multiple searches to use a different action

The user should 'search a search term' then 'search a different search term'