Add new properties to `Trust` and move creation to the helper class
When adding new properties to `Trust` all the unit tests broke! Have moved the creation of `Trust` objects in tests to a `DummyTrustFactory` so that adding new properties in the future is a lot less painful.
Creating a `Trust` from a `Group` has moved to a helper class in order to share it with the fake data generator