Update search pageObject model to use fake data
Allow json imports in typescript and add FakeData interface
Add import for fake data json and update assertion for `toSeeInformationForEachResult` to calculate the expected result.

This assertion also calculates how many results should be displayed on a page, using the upper limit for pagination (e.g. if pagination count is 20, then the max number of items on the page should be 20, but this count may be lower if there are less results).

Add trusts.json