Remove wiremock npm scripts
Previously it was an option to run tests locally against a wiremock container, to congigure the mock API dependency. Now, we will need to use only our custom docker compose script to setup the fake db. A developer can still run tests locally, by changing the connection string for the database to point to the one set up during the build of this docker compose file (as referenced in the GitHub docs).