Store the current trust on trust pages
When directly testing the trust and details pages for trusts, we will navigate to a specific trust and check the right information is shown. Storing this expected trust object on navigation to the page to simplify our tests.

We will need different assertions, such as 'toBeOnTheRightPageFor' to test navigation from different pages, as here we are not storing the current trust object.