Fix deployment and integration tests to match test updates
We updated the page object models significantly in #202 but as we were skipping integration and deployment tests, these were not updated. Updated these with new method names.

Made 'FakeTestData' dependency injection on a 'SearchPage' optional, as we will not be testing against fake test data in integration and deployment tests.