Create one parent faker class
As we will have multiple fakers, we need to have a method of easily adding a new faker.

Here we create one AcademiesDbFaker class which is responsible for creating and a data class containing generated data matching different tables in the Academies database.

Also updated the Json and Sql generators to take this class.

Updated Json generator to use the new chained trust builder methods to create the expected trust data for tests.

Updated Sql generator to be more type generic, and create one insert script for multiple tables.