Add page title for academies pages
Academies page (browser) titles are slightly different to the h1 shown on the page, so adding an optional property for Trust pages: `PageTitle`. If this is not set, the `PageName` property will be used to created the title.

Also added summaries to the properties in `TrustAreaModel` to make it clearer what each property is used for.