Create shared TrustsAreaModel for all trust pages
As all trust pages share the same data model we now have a lot of repeated code. This moves the majority of page functionality to a shared class, and only sets the page specific properties (such as the page name).

This means we can reduce duplicated code and tests when making changes.