Change all trust pages to use named query parameter
All urls will now display the Trust UID in the url as a named parameter: `?uid={UID}`. This change has been made for the trust details page, and we need to do it for the other pages in order to use the same Base Page Object Model in UI tests.