Move opens in new tab text out of table
As we have multiple links which open in a new tab in the table, we do not need to add '(opens in a new tab)' text next to every link, and can have it above the section containing multiple links. This was previously in the Get information about schools column header, but upon discussion with the design community have agreed that this should be in plain text content above the table containing links which open in a new tab.