Add (opens in new tab) to accessible link text
As per [GOV.UK guidance](https://design-system.service.gov.uk/styles/links/#opening-links-in-a-new-tab) if having one text reference to multiple links opening in a new tab, the link should include hidden text to tell screen reader users that the link in focus opens in a new tab (for when navigating by links).

As we are using an aria-label to read a different message for the link text (as 'More information' makes no sense out of context), adding this to the aria-label.