Make `mstrTrust` parameter optional when creating a `Trust`
Not all SATs and MATs in the `GIAS.Group` table have a corresponding entry in the `mstr.Trust` table. Currently there are 131 closed SATs (presumably converted to MATs) which are not in the `mstr.Trust` table. We want to be able to display as much information as possible so we have chosen to make the `mstrTrust` parameter optional