Fix MOJ sort icon causing content to overflow
The MOJ sort icon is positioned using absolute positioning, with a slight negative position on the container (`right: -1px`), which means if the last column is sortable, the table will always slightly overflow the parent container. This means that when using an `overflow:auto` property on the container to ensure the table does not exceed the width of the page on smaller screens, the horizontal scrollbar will always appear–even if it does not appear to be overflowing.

Have provided a slight tweak to the sort icons when used within this container:
- remove negative right-hand positioning
- add additional padding on the right hand side of the button to replace the spacing between title and icon provided by the negative positioning.