Increase stryker mutation thresholds in CI
Previously our CI was failing at a score of 80, but this has allowed files to be merged with much lower scores of 66, while the whole project score is 98. This should flag if we have any low scoring files we are trying to merge in. We will monitor if having a minimum score of 100 causes any problems.