Add 'not yet inspected' for all cells with no Ofsted inspection
Previously our design showed no value for a previous Ofsted rating, if the current Ofsted rating does not yet exist. However, we have updated our business logic to always show the 'Not yet inspected' value for both columns.

Adding this value in for the academy row with no Ofsted ratings, and updated the UI tests to check for both displaying 'Not yet inspected'.