Update Academies Db Faker to create establishments and grouplinks
Thi required changing the use of `Select` on the Trusts to generate list - as there are multiple side-effects of this loop. Academies are dependent on the same object, and we need to create a Group Link for each trust and set of academies. MstrTrusts, Group links and Academies are all dependent on the Group.

Have also changed the AcademiesDb data object from a record to a class.
This means we can update the lists for each model directly on the class rather than intatiating lists as variables.