Create one ref date for all fakers
We need to set a ref date when using any random dates in Faker, to ensure it doesn't gnerated a past or future date from the current date every time it is run (thus creating a new set of data). Setting this in one place as we only need to declare it once.