Hide links if no academies
Some links to other services will not exist if a trust is a Single Academy Trust and has no academies, so wrapping these in a boolean.

Remove the fallback empty strings for the methods to fetch the right link - as the new boolean will ensure the method is only called if a Trust is a Multi or single academy trust.