Add links to other services
Create a class which will store and build links to other services with the required query parameter. I want to separate this from the page logic to ensure that links that may need to be updated in the future are easy and obvious to change.

Showing a different link for single and multi-academy trusts for School Financial benchmarking and Find school performance links.

Delegate responsibility for building links for different establishments to the link builder. Set properties on the DetailsPage model to be nullable strings on page load - and check for null to determine whether to render the string, rathern than checking for specific scenarios in the view.

Some links to other services will not exist if a trust is a Single Academy Trust and has no academies, so adding a null check for each link - the link builder will return null if the link should not exist.

Add tests for OtherServicesLinkBuilder

Add unit tests for new link builder class

Update Dummy Academy and Trust factory to make it more flexible to create dummy trusts with different scenarios