Add links to other services
Create a class which will store and build links to other services with the required query parameter. I want to separate this from the page logic to ensure that links that may need to be updated in the future are easy and obvious to change.

Showing a different link for single and multi-academy trusts for School Financial benchmarking and Find school performance links.