Remove data volume when running `docker:stop`
We are using a shared volume when running the Faker, but this will not be overriden by the `FakeDataGenerator.Dockerfile` in between running the `docker-compose.ci.yml` file locally. Here adding a -v flag to `docker compose down` in our npm script to stop and restart the docker containers, but would be good to look at a way to add this to the docker-compose file itself, to ensure the docker-compose file uses a clean named volume.