Move DummyFactories to data test project
When we created the Dummy Factory pattern we did not have a test project for the Data project, as it was purely holding POCOs.

Now that we have tests for this project, it makes more sense to place these Mocks here, as they are more closely related to the models in this project.

- Removed the reference to the web app test project from the Data unit tests project
- Added a reference to the Data unit tests project from the Web app test project - so that we can continue to use the Dummy Factories in our web app unit tests.
- Replace reference in Academies DB unit tests project
- Update namespaces for Mocks - These should be the new project