Add UI tests for pupil numbers page
Test that the right data is shown on each row of the academies pupil numbers table, and that the right calculation is made for % full

Update `page.goTo()` method for academies pages (via `BaseAcademiesPage`) to ensure it goes to a page with academies. We will have a future user story which deals with the 0 academies scenario.