Default UI tests to state where cookies are already accepted
The cookies banner was causing some odd issues in the UI tests (such as unrelated buttons not working) which couldn't be replicated manually. We're circumventing these problems by using a storage state containing a consent cookie which never expires. This storage state is overridden in the cookies tests where we want a clean slate