Add establishment type matcher
Create a switch expression which matches the Gias establishment to the establishment name on Explore education statistics free school meals data.

This will be used to get the appropriate average for an an academy.

Change return type for provider methods to be non-nullable
The percentage will always have a value