Add static data class with hardcoded values
This stores the dictionary with Free school meals percentages, and populates it using a static constructor. It allows us to add a layer of separation from our provider, and makes it easier to test. This will also make it easier to change when the hardcoded data is replaced by DB or API layer.

We have only included the relevant establishment types for the Academies DB data (using the 4 keys on the `PhaseTypeGroup` enum).

The data also directly references the enum property rather than a string, removing the need for the 'Add' method on the POCO model.

Update `GetLaAverage` method to find the correct La in the data dictionary using the academy LaCode now that the data is populated.

This also enables us to add unit tests for FreeSchoolMeals provider:

- Add new test project for `~Data.Hardcoded` project
- Add tests for each method on data using a subset of hardcoded data for expected value
- Add tests for Free school meals data class

Add intellisense comments for Free school meals data class