Improve flakey test
This test was occasionally failing in CI, possibly because it was asserting the value of the input on the page before the page has reloaded.

Swapping the order of the assertions to check for the heading test before the component which is reliant on JavaScript.

Also reducing the number of times the user searches, as we only need to do that once in this test, and the beforeEach is already going to a page with a search term entered.