Change workflow to take advantage of Sonar now scanning non-dotnet code too
In a recent external release, Sonar scanner for dotnet changed from only scanning dotnet code (and code directly referenced by the csprojs) to scanning all code in the repo for all languages it supports (such as Docker and TypeScript) by default. This change to the workflow is to better describe that behaviour and to trigger by default unless the path is explicitly something that isn't scanned by Sonar.