Fix System.InvalidOperationException on Ofsted ratings page
`System.InvalidOperationException: The client projection contains a reference to a constant expression of 'DfE.FindInformationAcademiesTrusts.Data.OfstedRating'` was being thrown by LINQ trying to convert EF results to an anonymous type so added a concrete type for it to use. Changed return of `GetOfstedRatings` to also use this type for greater clarity over the previous Tuple.