Introduce new FIAT-specific data source repository
As pieces of data such as contact details are likely to be updated
individually rather than in bulk, the FIAT data source logic needs to
know about the specific thing being queried. The FIAT data source
repository provides a method to get a school contact by URN and contact
role, allowing the correct information to be provided on a per-item
basis.