Fix `InvalidOperationException` when no trust data exists for school
The `SchoolRepository.GetDateJoinedTrustAsync` method uses `FirstAsync`
in its query, which throws an `InvalidOperationException` when the query
returns an empty set of results.

Simply using `FirstOrDefaultAsync` instead isn't enough, as a `DateOnly`
is a value type, so the default value is `01-01-0001`, which ends up
being displayed on the page.

By calling `Cast<DateOnly?>` first, the default value becomes `null` (as
it's now a `Nullable<DateOnly>`), which can be handled appropriately in
the view.