Refactor `SchoolNavMenu` to be a service instead of a static class
Changes to the school pages will mean that the contents of the nav menu
will depend on a feature flag. Getting the `IVariantFeatureManager`
neatly into the `SchoolNavMenu` class is somewhat tricky - it's possible
to have `SchoolAreaModel` and its inheritors require one in their
constructors, but then many pages that won't actually need to know about
feature management will depend on one. An additional problem is that the
`SchoolNavMenu` exposes a synchronous interface, but the feature
management methods are `async`.

To make this easier, this commit refactors `SchoolNavMenu` by:
- making the class non-static, and introducing an `ISchoolNavMenu`
  interface that declares the two public methods, which the class
  implements
- registers the interface and class with the DI container so that it can
  be injected into the page models
- Updates the `SchoolAreaModel` class and its inheritors to have a
  dependency on the `ISchoolNavMenu` service, and to expose a
  `ServiceNavLinks` property and a `SubNavLinks` property
- Decouples the `_SchoolLayout.cshtml` partial from the `SchoolNavMenu`
  class by using the new properties

When feature management is added to the `SchoolNavMenu`, it will be
relatively straightforward to add a new dependency on the
`IVariantFeatureManager` and to modify the methods to be `async`, due to
the reduced coupling.