Return trust contacts as part of internal contacts for an academy
This updates the `SchoolContactsService.GetInternalContactsAsync` method
to fetch contact information from a school's trust, if it has one. Like
the trust itself, an academy should have the trust relationship manager
and the SFSO lead as contacts.

The `SchoolInternalContactsServiceModel` record has also been updated to
be able to hold the trust relationship manager and SFSO lead
information.