Add academy-specific contacts to 'In Dfe' page
This adds the trust relationship manager and SFSO lead to the 'Contacts
in DfE' page when the school being viewed is an academy. These contacts
will not be displayed if the school is maintained by the local
authority; likewise, the regions group local authority lead will no
longer be displayed if the school is an academy.